feat: add user delete handler
