+++ src/controller/users/deleteRequestHandlers.ts
@@ -1,0 +1,6 @@
+import { UUID } from 'crypto';
+import { db } from 'src/main';
+
+export function removeUserFromDb(id: UUID) {
+  db.removeUserFromDb(id);
+}
